fixes for failing tests
